requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==21.4.0
black==22.1.0
cfgv==3.3.1
chardet==3.0.4
click==8.0.3
discord.py==2.0.0a3686+g1fae5bf
distlib==0.3.4
filelock==3.4.2
identify==2.4.7
idna==3.3
isort==5.10.1
libcst==0.3.23
multidict==4.7.6
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.1
pre-commit==2.17.0
pycln==1.1.0
python-dotenv==0.19.2
PyYAML==5.4.1
six==1.16.0
toml==0.10.2
tomli==2.0.0
typer==0.4.0
typing-inspect==0.7.1
typing_extensions==4.0.1
virtualenv==20.13.0
yarl==1.5.1
